refactor: separate workflow ports from domain values
