+++ tools/dev-workflow-v2/src/features/workflow/domain/git-status.ts
@@ -1,0 +1,8 @@
+/** @riviere-role value-object */
+export interface WorkflowGitStatus {
+  readonly changedFilesVsDefault: readonly string[]
+  readonly currentBranch: string
+  readonly hasCommitsVsDefault: boolean
+  readonly headCommit: string
+  readonly workingTreeClean: boolean
+}
