+++ tools/dev-workflow-v2/src/features/workflow/domain/ports/create-pull-request.ts
@@ -1,0 +1,5 @@
+import type { PullRequestCreationRequest } from '../pull-request-description'
+import type { WorkflowPullRequest } from '../pull-request'
+
+/** @riviere-role domain-port */
+export type CreateWorkflowPullRequest = (request: PullRequestCreationRequest) => WorkflowPullRequest
