+++ tools/dev-workflow-v2/src/features/workflow/domain/ports/read-git-status.ts
@@ -1,0 +1,4 @@
+import type { WorkflowGitStatus } from '../git-status'
+
+/** @riviere-role domain-port */
+export type ReadWorkflowGitStatus = () => WorkflowGitStatus
