+++ tools/dev-workflow-v2/src/features/workflow/domain/ports/read-pull-request-feedback.ts
@@ -1,0 +1,4 @@
+import type { WorkflowPullRequestFeedback } from '../pull-request-feedback'
+
+/** @riviere-role domain-port */
+export type ReadWorkflowPullRequestFeedback = (prNumber: number) => WorkflowPullRequestFeedback
