+++ tools/dev-workflow-v2/src/features/workflow/domain/pull-request.ts
@@ -1,0 +1,6 @@
+/** @riviere-role value-object */
+export interface WorkflowPullRequest {
+  readonly prNumber: number
+  readonly prUrl: string
+  readonly isDraft: boolean
+}
